Add Animator, View, Draw Functions

+++ source/Engine/Scene.cpp
@@ -3572,7 +3572,7 @@ PUBLIC STATIC bool Scene::ObjectTileCollision(Entity* entity, int cLayers, int c
     }
 }
 
-PUBLIC STATIC bool Scene::ObjectTileGrip(Entity* entity, int cLayers, int cMode, int cPlane, float xOffset, float yOffset, float tolerance) {
+PUBLIC STATIC bool Scene::ObjectTileGrip(Entity* entity, int cLayers, int cMode, int cPlane, int xOffset, int yOffset, float tolerance) {
     int layerID     = 1;
     bool collided   = false;
     int posX        = xOffset + entity->X;
